feat(voice): Rust语音模块完整重构，超越C++实现
核心改进：

- 新增高级功能：GCC带宽估计、AEC回声消除、AGC2自动增益、MOS评分

- 性能优化：热路径零分配、无锁队列、位运算优化、Worker线程

- 安全加固：FFI边界保护、协议解析检查、null指针验证、RFC1982序列号

- 测试覆盖：280+测试用例、7种信号类型、性能回归测试

- 文档完善：GCC/AEC/AGC算法文档、参考文献

修改统计：47文件，+9684/-1775行

+++ src/rust-bridge/voice/src/dsp/gain.rs
@@ -3,11 +3,11 @@
 //! 支持 0-300% 增益，超过 100% 时进行硬削波
 
 /// 应用增益到音频样本
-/// 
+///
 /// # 参数
 /// * `gain` - 增益倍数 (0.0 - 3.0)
 /// * `samples` - 音频样本数组 (会被原地修改)
-/// 
+///
 /// # 示例
 /// ```
 /// use ddnet_voice::dsp::gain;
@@ -16,7 +16,7 @@
 /// ```
 pub fn apply(gain: f32, samples: &mut [i16]) {
     let gain = gain.clamp(0.0, 3.0);
-    
+
     for sample in samples.iter_mut() {
         let amplified = *sample as f32 * gain;
         *sample = amplified.clamp(-32768.0, 32767.0) as i16;
@@ -28,13 +28,10 @@ pub fn frame_peak(samples: &[i16]) -> f32 {
     if samples.is_empty() {
         return 0.0;
     }
-    
-    let max_sample = samples
-        .iter()
-        .map(|&s| s.abs())
-        .max()
-        .unwrap_or(0) as f32;
-    
+
+    // 使用 i32 避免 i16::MIN.abs() 溢出问题
+    let max_sample = samples.iter().map(|&s| (s as i32).abs()).max().unwrap_or(0) as f32;
+
     max_sample / 32768.0
 }
 
@@ -43,12 +40,9 @@ pub fn frame_rms(samples: &[i16]) -> f32 {
     if samples.is_empty() {
         return 0.0;
     }
-    
-    let sum_squares: f64 = samples
-        .iter()
-        .map(|&s| (s as f64).powi(2))
-        .sum();
-    
+
+    let sum_squares: f64 = samples.iter().map(|&s| (s as f64).powi(2)).sum();
+
     let rms = (sum_squares / samples.len() as f64).sqrt() as f32;
     rms / 32768.0
 }
@@ -61,7 +55,7 @@ mod tests {
     fn test_gain_apply() {
         let mut samples = vec![1000i16; 960];
         apply(2.0, &mut samples);
-        
+
         // 增益 200% 应该使值翻倍
         for sample in samples {
             assert_eq!(sample, 2000);
@@ -72,7 +66,7 @@ mod tests {
     fn test_gain_clamp() {
         let mut samples = vec![20000i16; 960];
         apply(3.0, &mut samples); // 300% 增益
-        
+
         // 应该被限制在 i16 范围内
         for sample in samples {
             assert!(sample <= 32767);
@@ -84,7 +78,7 @@ mod tests {
     fn test_gain_zero() {
         let mut samples = vec![1000i16; 960];
         apply(0.0, &mut samples);
-        
+
         // 0% 增益应该全部归零
         for sample in samples {
             assert_eq!(sample, 0);
